refactor(RepoWidget): Extract mapParser function
